[CSL-1171] Split Delegaton.Logic
