+++ cardano-sl.cabal
@@ -205,6 +205,9 @@ library
                         Pos.Delegation.Helpers
                         Pos.Delegation.Listeners
                         Pos.Delegation.Logic
+                        Pos.Delegation.Logic.Common
+                        Pos.Delegation.Logic.Mempool
+                        Pos.Delegation.Logic.VAR
                         Pos.Delegation.Lrc
                         Pos.Delegation.Types
                         Pos.Delegation.Worker
